requirements.txt
Django>=2.2.8,<3.2
djangorestframework>=3.9.0,<3.12.2
flake8>=3.0,<3.9.1
psycopg2>=2.2,<2.8.6   
